refactoring entire package regarding pyi files (for AI and IDE)
