+++ bittensor_wallet/config/__init__.py
@@ -1,0 +1,3 @@
+from bittensor_wallet.bittensor_wallet import config as _
+
+Config = _.Config
