+++ bittensor_wallet/errors/__init__.py
@@ -1,7 +1,6 @@
-from typing import TypeAlias
-
 from bittensor_wallet.bittensor_wallet import errors as _
 
-ConfigurationError: TypeAlias = _.ConfigurationError
-KeyFileError: TypeAlias = _.KeyFileError
-PasswordError: TypeAlias = _.PasswordError
+ConfigurationError = _.ConfigurationError
+KeyFileError = _.KeyFileError
+PasswordError = _.PasswordError
+WalletError = _.WalletError
