+++ bittensor_wallet/errors/__init__.pyi
@@ -1,0 +1,7 @@
+class KeyFileError(Exception): ...
+
+class ConfigurationError(Exception): ...
+
+class PasswordError(Exception): ...
+
+class WalletError(Exception): ...
