+++ bittensor_wallet/keypair/__init__.py
@@ -1,5 +1,3 @@
-from typing import TypeAlias
-
 from bittensor_wallet.bittensor_wallet import keypair as _
 
-Keypair: TypeAlias = _.Keypair
+Keypair = _.Keypair
