+++ bittensor_wallet/utils/__init__.pyi
@@ -1,0 +1,12 @@
+from typing import Final, Union
+
+
+SS58_FORMAT: Final[int]
+
+def is_valid_ss58_address(address: str) -> bool: ...
+
+def get_ss58_format(ss58_address: str) -> int: ...
+
+def is_valid_ed25519_pubkey(public_key: Union[str, bytes]) -> bool: ...
+
+def is_valid_bittensor_address_or_public_key(address: Union[str, bytes]) -> bool: ...
