+++ bittensor_wallet/wallet/__init__.py
@@ -1,6 +1,4 @@
-from typing import TypeAlias
-
 from bittensor_wallet.bittensor_wallet import wallet as _
 
 display_mnemonic_msg = _.display_mnemonic_msg
-Wallet: TypeAlias = _.Wallet
+Wallet = _.Wallet
